refactor(qwen3vl-embedding): unify embedding runtime under utils
